requirements.txt
streamlit
python-dotenv
groq
pandas
numpy
plotly


